pyproject.toml
[tool.poetry]
name = "pythia"
version = "1.0.0"
description = ""
authors = ["Beda Kosata <beda@bigdigdata.com>"]

[tool.poetry.dependencies]
"ufal.morphodita" = "~1.10.1.1"
anytree = "^2.8.0"
django = "^4.2.16"
django-allauth = "~64.2.1"
django-anymail = "^11.1"
django-cachalot = "~2.6.3"
django-compressor = "~4.5.1"
django-js-asset = "~2.2.0"
django-mptt = "~0.14.0"
django-redis = "~5.4.0"
djangorestframework = "~3.15.2"
drf-nested-routers = "~0.94.1"
elasticsearch = "~8.15.0"
factory-boy = "^3.3.1"
faker = "^28.4.1"
fastcache = "^1.1.0"
# gensim = "~4.2.0"
isbnlib = "^3.10.10"
lz4 = "~4.3.3"
nltk = "^3.9.1"
oaiharvest = "~3.0.0"
psycopg2-binary = "~2.9.3"
pyfakefs = "^5.6.0"
pymarc = "~5.2.2"
pyoai = "2.5.0"  # 2.5.1 cannot be installed for some reason
pysftp = "^0.2.9"
python = ">=3.8,<4.0"
python-dateutil = "~2.8.1"
python-decouple = "^3.3"
pytz = "~2024.1"
redis = "~5.0.8"
requests = "~2.32.3"
tabulate = "~0.9.0"
tqdm = "~4.66.5"
xmltodict = "^0.13.0"
zstandard = "^0.23.0"
extmaillogin = "0.4.0"
dj-rest-auth = "~6.0.0"

[tool.poetry.dev-dependencies]
black = "24.8.0"
django-debug-toolbar = "^4.4.6"
django-extensions = "^3.2.0"
django-querycount = "^0.8.3"
# pandas = "^1.4.3"
pre-commit = "^3.4.0"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
pytest-django = "^4.9.0"

[tool.black]
line-length = 100
skip-string-normalization = true
target-version = ["py36"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
